Handle projects with custom iOS app delegate, by commenting the App Delegate directive in the SingularAppDelegate.m file
