+++ SingularSDK/Editor/SingularEditorParams.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 58e0c8cb5925438790ea3a52154a5c20
+timeCreated: 1720604858
